+++ test/unit/configuration.unit.test.ts
@@ -8,7 +8,7 @@ describe("Configuration unit tests suite", () => {
   });
 
   it("Can add token config of unknown token", () => {
-    assert.throws(() => configuration.tokens.getDecimals("UnknownToken"));
+    assert.equal(configuration.tokens.getDecimals("UnknownToken"), undefined);
 
     configuration.updateConfiguration({
       tokens: {
@@ -22,8 +22,8 @@ describe("Configuration unit tests suite", () => {
   });
 
   it("Adding token config does not affect existing config", () => {
-    assert.throws(() => configuration.tokens.getDecimals("UnknownToken1"));
-    assert.throws(() => configuration.tokens.getDecimals("UnknownToken2"));
+    assert.equal(configuration.tokens.getDecimals("UnknownToken1"), undefined);
+    assert.equal(configuration.tokens.getDecimals("UnknownToken2"), undefined);
 
     configuration.updateConfiguration({
       tokens: {
@@ -49,7 +49,7 @@ describe("Configuration unit tests suite", () => {
 
   it("Reset of configuration reverts additions and changes", () => {
     assert.equal(configuration.tokens.getDecimals("TokenA"), 18);
-    assert.throws(() => configuration.tokens.getDecimals("UnknownToken"));
+    assert.equal(configuration.tokens.getDecimals("UnknownToken"), undefined);
 
     configuration.updateConfiguration({
       tokens: {
@@ -68,7 +68,7 @@ describe("Configuration unit tests suite", () => {
     configuration.resetConfiguration();
 
     assert.equal(configuration.tokens.getDecimals("TokenA"), 18);
-    assert.throws(() => configuration.tokens.getDecimals("UnknownToken"));
+    assert.equal(configuration.tokens.getDecimals("UnknownToken"), undefined);
   });
 
   it("can read mangroveOrder config", () => {