+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Upgrade `examples/tutorials/on-the-fly-offer.js` to new Mangrove core protocol and SDK
 - fix: Coerce ticks to tickSpacing when given as arguments
 - feat: Add integration test of tickSpacing>1
+- feat!: `configuration.tokensConfiguration.getDecimals` now returns `undefined` instead of throwing if the token decimals are not in the configuration
 
 # 2.0.4